requirements-research.txt
beautifulsoup4
requests
pyyaml
tqdm
numpy
torch
transformers
datasets
seqeval
faiss-cpu
pytest
